backend/requirements.txt
fastapi
uvicorn[standard]
pydantic[email]
python-dotenv
numpy
pandas
scikit-learn
xgboost
shap
reportlab
fpdf
requests
passlib[bcrypt]
python-jose[cryptography]
